fix : java sh04 ex04 change int to float in constructors

+++ Java/sh04/src/ex04/FreshProduct.java
@@ -6,7 +6,7 @@ public class FreshProduct extends Product{
     public FreshProduct() {
     }
 
-    public FreshProduct(String name, int price, int daysOfExpiration) {
+    public FreshProduct(String name, float price, int daysOfExpiration) {
         super(name, price);
         this.daysOfExpiration = daysOfExpiration;
     }
